test: set build constraints for Gont-based tests
