--- pkg/core/core_test.go
@@ -1,27 +1,0 @@
-package core_test
-
-import (
-	"math/rand"
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-
-	"riasc.eu/wice/pkg/util"
-	"riasc.eu/wice/test"
-)
-
-func TestSuite(t *testing.T) {
-	rand.Seed(GinkgoRandomSeed())
-
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Core Suite")
-}
-
-var _ = test.SetupLogging()
-
-var _ = BeforeSuite(func() {
-	if !util.HasAdminPrivileges() {
-		Skip("Insufficient privileges")
-	}
-})
